Decompose test bulk/search helper. Fix flaky test - bulk refresh true.
